chore(TCOMP-681): move TaCoKitComposite to correct package
