+++ component-studio-integration/src/main/java/org/talend/sdk/component/studio/ui/composite/TaCoKitComposite.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.talend.sdk.component.studio.view;
+package org.talend.sdk.component.studio.ui.composite;
 
 import static org.talend.sdk.component.studio.metadata.ITaCoKitElementParameterEventProperties.EVENT_PROPERTY_VALUE_CHANGED;
